+++ yandex/data_source_yandex_mdb_postgresql_cluster.go
@@ -36,6 +36,10 @@ func dataSourceYandexMDBPostgreSQLCluster() *schema.Resource {
 										Type:     schema.TypeBool,
 										Computed: true,
 									},
+									"serverless": {
+										Type:     schema.TypeBool,
+										Computed: true,
+									},
 								},
 							},
 						},